fixed the health data to be floats

+++ Code/createDatabase.sql
@@ -1,10 +1,10 @@
 CREATE TABLE Country(
-    countryName VARCHAR(50), 
-    numCases int, 
+    countryName VARCHAR(50),
+    numCases int,
     numDeaths int,
     numRecovered int,
-    numDoctors int,
-    numHospitalBeds int,
+    numDoctors float(5),
+    numHospitalBeds float(5),
     latestTravelRestriction VARCHAR(2000),
     PRIMARY KEY (countryName)
 );
